Apply pnacl irt shim change to the generator, not the generated file.
Cast to int to avoid overflow.

BUG=none


Review URL: https://chromiumcodereview.appspot.com/12315107

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@184726 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: b8b7debd205af85a1214e5cb907bea7549bdbc42

+++ ppapi/native_client/src/untrusted/pnacl_irt_shim/pnacl_shim.c
@@ -121,7 +121,7 @@ static int mystrcmp(const char* s1, const char *s2) {
     ++s1;
     ++s2;
   }
-  return *(s1) - *(s2);
+  return (int)(*s1) - (int)(*s2);
 }
 
 /* BEGIN Declarations for all Wrapper Infos */
